#93: home carer, report history 수정
