+++ lib/main.dart
@@ -8,7 +8,7 @@ import 'package:atti/screen/LogInSignUp/SignUpScreen2.dart';
 import 'package:atti/screen/LogInSignUp/SignUpScreen3.dart';
 import 'package:atti/screen/LoginSignUp/IntroPage.dart';
 import 'package:atti/screen/Notice/NoticeMain.dart';
-import 'package:atti/screen/report/ReportDetail.dart';
+import 'package:atti/screen/report/_ReportDetail.dart';
 import 'package:atti/screen/report/ReportHistory.dart';
 import 'package:atti/screen/routine/RoutineMain.dart';
 import 'package:atti/screen/routine/register/RoutineRegister1.dart';
